GUI Components -> Rust
index.html just holds the script tag in the body now. We do a nasty hack
in Percy to workaround the lack of an `Rc<VirtualNode>` embedding. See
https://github.com/chinedufn/percy/issues/108

We suffer with an O(n^2) virtual node embedding for now. It seems like
with the current WASM work, we only spend 0.85ms in our Rust code! We
have plenty of time to spare for actual GameBoy logic.

Included here is an unused (for now) memory table debug interface that
will eventually be hooked into the data and GUI.